- fix matrix inverse for row-major simple matrixes; - repair some tests and add better coverage - many widespread small formatting and style fixes
apologies for the terrible checkin hygiene to my future self and any other poor schmuck who might have to figure out what the hell happened here. this branch is a mess and I just need to square it away and get back to master. hopefully this kind of thing should not happen again.